hot fix

+++ src/gpuworker/kernel32.cl
@@ -93,8 +93,8 @@ static int is_live_in_sieve9(uint128_t n)
 	ulong r = 0;
 
 	r += (uint)(n);
-	r += (uint)(n >> 32) * 4;
-	r += (uint)(n >> 64) * 7;
+	r += (uint)(n >> 32) * (ulong)4;
+	r += (uint)(n >> 64) * (ulong)7;
 	r += (uint)(n >> 96);
 
 	r = r % 9;
